requirements.txt
streamlit==1.20.0
pandas==1.5.3
scikit-learn==1.2.2
requests==2.28.2
beautifulsoup4==4.12.2
google-generativeai==0.7.2
selenium==4.10.0
flask==2.3.2